Activating TALC's HPC module environment within Notebooks
The list of modules is automatically activated as the first
cell of each Jupyter Notebook assignment file.

Furthermore, a caution statement is added to warn users to activate
the appropriate Pyhton enviornment (or Kernel) before running
the Notebooks.

Also, the path to the SUMMA executable is modified. Users do
not need to modify the path to run the Notebooks.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>