Adding instruction to load necessary modules on TALC UCalgary
The list of modules need to be activated whenever users would like
to use the workflows for this course. The instructions and relevant
workflows and information to load the modules are added.

The HPC modules are automatically loaded in Jupyter Notebooks of the
assignments. Users only need to run the relevant Notebook cells.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>