Bypassing usage of HydroShare API's Python package
Based on a recent conversation with Clara Cogswell, HydroShare
team is "working on streamlining [their] API". Therefore, the
method to download the resource has been changed to directly
use the HTTPS request.

Furthermore, when student would like to run the `2c` simulations
themselves, they may face a "Permisison denied" error. To avoid
this, another cell is added to assure the downloaded NetCDF
file are renamed to have the "_backup" suffix at the end, so
student do not face this error, while keeping whatever has been
downloaded from HydroShare as a backup. A sentence has been
also added to inform students of the renaming procedure.

Reported-by: Martyn Clark <martyn.clark@ucalgary.ca>
Signef-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>